worker1/requirements.txt
requests
kafka-python==2.0.2



